preimage metadata; GetJAMNetwork dot+jam support (#436)
* added metadata to ListServices responses
* added GetJAMNetwork (dot vs jam) + GetJAMNetworkPort (9800 [prev 9900]) abstractions using JAM_NETWORK env var
* used epoch0timestamp to have: "jam" -- actual JCE vs "jamtestnet" -- slot 12
* added 30s context cancels in fib WP submission
* setup timeout=0 for long-running "fib" test to supports a  LONG running "fib" (like 1K-10K) WP submission **without** network resets every hour

```
while true; do make run mode=fib; sleep 1; done
```

* Misc: monitoring goroutine usage `processBlock` like
```
INFO [04-09|00:54:42.103] processBlock                             author=2 p=ae46..03ce h=3771..05b1 H_t=1404547 goroutines=71
```

+++ node/node_jamtest.go
@@ -122,7 +122,7 @@ func computeLevelDBPath(id string, unixtimestamp int, jobID string) (string, err
 	return path, nil
 }
 
-func SetupQuicNetwork(network string, basePort uint16) (uint32, []string, map[uint16]*Peer, []types.ValidatorSecret, []string, error) {
+func SetupQuicNetwork(network string, basePort uint16) (uint64, []string, map[uint16]*Peer, []types.ValidatorSecret, []string, error) {
 	peers := make([]string, numNodes)
 	peerList := make(map[uint16]*Peer)
 	nodePaths := SetLevelDBPaths(numNodes)
